+++ CHANGELOG.md
@@ -1,11 +1,18 @@
-## [1.5.2] - 09 November 2021
+## [1.5.3] - 18 November 2020
+
+- PaginatedGridView and PaginatedListView now can be use in another scrollView widget
+- Add DotTabIndicator
+- Other widget improvement
+- CustomBackButton widget improvement
+
+## [1.5.2] - 09 November 2020
 
 - Fix [JinAccordion issue](https://github.com/chunlee-thong/jin_widget_helper/issues/3)
 - add startIcon and endIcon to [JinLoadingButton][jinloadingbuttonurl]
 - add width and height to [JinLoadingButton][jinloadingbuttonurl]
 - CustomBackButton widget improvement
 
-## [1.5.1] - 02 November 2021
+## [1.5.1] - 02 November 2020
 
 - Extension improvement
 - improve StreamHandler and FutureHandler error widget